ap/new/requirements.txt
streamlit
numpy
joblib
openai-whisper
keybert
pytrends
pandas
scikit-learn
